Fixes user facing bugs
Found a lot of bugs with the editing and creating functions.
Need to find a way to only show buttons for edit and delete if the user
own the event.
Changed the back end so that users can only see their own events as a
temporary fix to this problem.
Still need to provide messages for each action.